feature_iep/requirements.txt
fastapi==0.110.0
uvicorn==0.30.0
torch==2.2.0
torchvision==0.17.0
opencv-python==4.9.0.80
numpy==1.26.3
python-multipart==0.0.9
pydantic==2.6.1
pillow==10.4.0 